refactor: extract filter state into useRestaurantFilters hook
